+++ terra-rust-cli/Changelog.md
@@ -1,6 +1,7 @@
 # Changes
 ## 1.0
-### 1.0.3 - Devel
+### 1.0.3 - 20-Feb-2022
+* get_json_block - reads stdin/file/or json passed as an argument
 ### 1.0.2 - 16-Feb-2022
 * error returns
 ### 1.0.1